requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
datasets>=2.14.0
faiss-cpu>=1.7.4
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
tiktoken>=0.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6
pydantic>=2.0.0
ragas>=0.1.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wandb>=0.16.0
python-dotenv>=1.0.0
pypdf>=3.17.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0